+++ pkg/providers/property/testdata/TestDataPropertyHostnamesDiff/valid_no_contract_and_group.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_hostnames_diff" "diff" {
+  property_id = "prp_1"
+}